Add 64 bit config saving (not needed for now tho')

+++ Core/Config.h
@@ -54,7 +54,7 @@ struct ConfigTouchPos {
 };
 
 struct ConfigCustomButton {
-	int key;
+	uint64_t key;
 	int image;
 	int shape;
 	float rotation;
